refactor shouldCompleteMultiline
